Fix #274 - parser.GetCustomData() throws FileNotFoundException for missing DACPACs.
The FileNotFoundException is OK if the reference of the root DACPAC is not found.
It is however not OK if the missing DACPAC is only referenced by another DACPAC than the root DACPAC, but not by the root DACPAC directly.
Therefore, skip missing DACPACs if they're not directly referenced by the root DACPAC.